requirements.txt
flask==3.0.3
flask-cors==5.0.0
requests==2.32.3
pandas==2.2.3
google-cloud-bigquery==3.26.0
google-auth==2.35.0
pyarrow==17.0.0
certifi==2024.8.30
google-generativeai==0.3.2
